simplify lib files
