+++ src/components/CodeExample.astro
@@ -1,7 +1,7 @@
 ---
 import { Code } from "astro-expressive-code/components";
 import { extractSection } from "@/lib/codeExamples/codeSnippets";
-import { loadRawFiles } from "@/lib/codeExamples/loadRawFiles";
+import { loadRawFiles } from "@/lib/codeExamples/loaders";
 
 const files = loadRawFiles();
