+++ src/components/CodeExampleGroup.astro
@@ -2,10 +2,8 @@
 import { Tabs, TabItem } from '@astrojs/starlight/components';
 import { Code } from 'astro-expressive-code/components';
 import { extractSection } from '@/lib/codeExamples/codeSnippets';
-import { resolveMeta } from '@/lib/codeExamples/resolveMeta';
-import { loadManifests } from '@/lib/codeExamples/loadManifests';
-import { loadRawFiles } from '@/lib/codeExamples/loadRawFiles';
-import { getGroupEntries } from '@/lib/codeExamples/manifestEntries';
+import { loadManifests, loadRawFiles } from '@/lib/codeExamples/loaders';
+import { resolveMeta, getGroupEntries } from '@/lib/codeExamples/manifest';
 import CodeExampleViewLink from './CodeExampleViewLink.astro';
 
 const rawFiles = loadRawFiles();
