+++ src/components/CodeExampleViewLink.astro
@@ -1,5 +1,5 @@
 ---
-import { loadManifests } from '@/lib/codeExamples/loadManifests';
+import { loadManifests } from '@/lib/codeExamples/loaders';
 
 interface Props { dir: string; group?: string; key?: string; section?: string; label?: string }
 const { dir, group, key, section, label } = Astro.props;
